Extract YouTubePlayerWebView.Coordinator into its own file
Resolves type_body_length SwiftLint violation introduced when the
autoplayArmer user-script registration pushed the struct over 350 lines.
Coordinator is now declared via an extension in
YouTubePlayerWebView+Coordinator.swift, no behavior change.
